pyproject.toml
[project]
name = "ai-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.5.1",
    "google-generativeai>=0.8.5",
    "pydantic-settings>=2.9.1",
    "pyyaml>=6.0.2",
]
